garment update request refactoring
